move Dhcp.* and Dns.* to library root
